# api/models/consume_product_stock_request_body.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// ConsumeProductStockRequestBody consume product stock request body
//
// swagger:model ConsumeProductStockRequestBody
type ConsumeProductStockRequestBody struct {

	// Quantity consumed
	// Required: true
	Quantity *int64 `json:"quantity"`
}

// Validate validates this consume product stock request body
func (m *ConsumeProductStockRequestBody) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateQuantity(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *ConsumeProductStockRequestBody) validateQuantity(formats strfmt.Registry) error {

	if err := validate.Required("quantity", "body", m.Quantity); err != nil {
		return err
	}

	return nil
}

// ContextValidate validates this consume product stock request body based on context it is used
func (m *ConsumeProductStockRequestBody) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *ConsumeProductStockRequestBody) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *ConsumeProductStockRequestBody) UnmarshalBinary(b []byte) error {
	var res ConsumeProductStockRequestBody
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}